put gym helper controller in controllers package
